src/requirements.txt
langchain==0.3.25
llama-cpp-python==0.3.9
huggingface-hub==0.32.2
langchain-chroma==0.2.4
chromadb==1.0.10
unstructured==0.17.2
python-telegram-bot==22.1